requirements.txt
streamlit==1.29.0
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
anthropic[vertex]==0.8.1
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
xlsxwriter==3.1.9
